+++ src/test/java/racingcar/util/RandomValueGeneratorTest.java
@@ -1,8 +1,7 @@
-package racingcar;
+package racingcar.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.util.RandomValueGenerator;
 
 import java.util.List;
